Create strategy-sushi-usdc-bct.sol
Here is a basic compounding strategy for usdc-bct pool on sushiswap Polygon. My first attempt, so hopefully it works. I wasn't sure what the poolId was or where to find it on Sushi, so I'm awaiting a response from Larry regarding that. I believe that we need MiniChef here because we aren't using an Eth pair, but again, not 100% sure.

The website for BCT is toucan.earth incase you were curious.